Add PMM.sci
Implements "Micromechanical formulas for the relaxation tensor of linear viscoelastic composites with transversely isotropic fibers", Barbero, E.J. and Luciano, R.,  International Journal of Solids and Structures, 32(13):1859--1872, 1995 for the particular case of linear elastic constituents. Provides a unified set of formulas for the five elastic properties of a transversely isotropic composite without fudge, adjustable, empirical, or any type of adjustable parameters. The formulas are analytical and closed form.